requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.1
python-multipart==0.0.6
PyPDF2==3.0.1
PyMuPDF>=1.24.0
Pillow>=10.0.0
pdf2docx>=0.5.9
psycopg2-binary==2.9.11
